requirements.txt
torch>=2.1.0
torch-geometric>=2.3.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
networkx>=3.0
accelerate>=0.25.0
sentencepiece>=0.1.99
matplotlib>=3.7.0
tqdm>=4.66.0
pyyaml>=6.0
paramiko>=3.3.0
protobuf>=3.20.0
fair-esm
biopython
tensorboard
tensorboardX
requests
pytest
seaborn>=0.13.0
